refactor: moves inline field context up to src dir
